Extract the statuses colors.
